requirements.txt
django==4.1.3
python-dotenv==0.20.0
psycopg2==2.9.5
djangorestframework==3.14.0
drf_spectacular==0.24.2